+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.1] / 2022-01-19
+- Remove not used file `Build`
+- Fix Version Information default value
+- GetValue add defaulValue
+- Find Changelog File
+- Update `IHazChangelog` 
+
 ## [1.0.0] / 2022-01-19
 - Update to public Repo
 
@@ -109,6 +116,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - First Release
 
 [vNext]: ../../compare/1.0.0...HEAD
+[1.0.1]: ../../compare/1.0.0...1.0.1
 [1.0.0]: ../../compare/0.0.16...1.0.0
 [0.0.16]: ../../compare/0.0.15...0.0.16
 [0.0.15]: ../../compare/0.0.14...0.0.15